# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskStatus = typeof TaskStatus[keyof typeof TaskStatus];


export const TaskStatus = {
  todo: 'todo',
  in_progress: 'in_progress',
  done: 'done',
} as const;

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Task {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  /** @nullable */
  dueDate: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  /** @nullable */
  description?: string | null;
  status?: TaskStatus;
  priority?: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
}

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: TaskStatus;
  priority?: TaskPriority;
  /** @nullable */
  dueDate?: string | null;
}

export interface DashboardSummary {
  total: number;
  todo: number;
  inProgress: number;
  done: number;
  dueToday: number;
  overdue: number;
  recentTasks: Task[];
}

export type ListTasksParams = {
status?: TaskStatus;
priority?: TaskPriority;
search?: string;
};