# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Task } from './task';

export interface DashboardSummary {
  total: number;
  todo: number;
  inProgress: number;
  done: number;
  dueToday: number;
  overdue: number;
  recentTasks: Task[];
}