# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from './taskPriority';
import type { TaskStatus } from './taskStatus';

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: TaskStatus;
  priority?: TaskPriority;
  /** @nullable */
  dueDate?: Date | null;
}